refactor(mouse): split scroll and forwarding out, completing the decomposition (F4 Part B, 3/n)
Final step, combining the two remaining clusters at the owner's request —
both small, mechanically identical verbatim moves that don't interact.

`scroll.rs` (182 production lines): wheel handling for a child that
ignores mouse reports — translate a tick into the agent's own verified
scroll keybinding, escalate a sustained streak to page keys, drive
codex's `^T` overlay.

`forward.rs` (193): forwarding to a mouse-aware child, the focus entry
points, and `mouse_report`'s coordinate translation.

`src/app/mouse.rs` was 1,529 production lines when this started — the
largest file in the tree, roughly twice the convention. The result:

    mod.rs        307   data types + handle_mouse (the one dispatch entry)
    route.rs      526   the pure decision layer
    selection.rs  399   the four begin/extend/finish triples
    forward.rs    193
    scroll.rs     182

Visibility cost across all three steps stayed proportionate: one struct's
fields in step 1, twelve `pub(super)` methods in step 2, eight here —
each one an entry point `handle_mouse` actually dispatches. No helper
had to leak, which is the evidence these are real seams and not
line-count cuts.

No behaviour change. 1,788 tests pass.
