+++ AGENTS.md
@@ -57,7 +57,7 @@ Deep design decisions live in [`ARCHITECTURE.md`](ARCHITECTURE.md) (sync-only, M
 - Event-loop machinery — **`sources.rs`** (channel coalesce + ingest), **`loop_steps.rs`** (pre-recv drain/refresh; chord-hint settle), **`streaming.rs`** (pull-source drains; stamps `last_output_at`), **`pane_wake.rs`**, **`scheduler.rs`** (timers/deadlines incl. agent idle/anim), **`watch.rs`** (off-thread fs-watch worker).
 - Pager — **`pager_handler/`** (vi-key handling, open/close hub, modal image overlay), **`pager_stream.rs`** + **`grep_session.rs`** + **`git_view_session.rs`** (the "worker → waking channel → streaming pager" abstraction; all ride one `stream_id` / `Message::PagerStreamOutput`), **`pager_history.rs`**.
 - Pure decisions (snapshot + fn + tests) — **`route.rs`** (`route_input` → `InputSink`, shared by key + paste), **`modal.rs`** (the transient `Modal` axis), **`focus.rs`** (`decide_focus` / `recompute_focus`).
-- **`mouse/`** — mouse routing. **`route.rs`** is the pure half: a `Copy` `MouseSnapshot` + pure `route_mouse` → `MouseSink`, plus `region_at` hit-testing the pointer against `compute_layout`'s `FrameLayout` — **the pointer, never keyboard focus**, which is what makes the wheel scroll what you're looking at. **`selection.rs`** holds the four drag-select clusters (pager text / chrome / pane text / list rows), each a `begin`+`extend`+`finish` triple where the press *claims* the drag and an unclaimed press leaves child-forwarding untouched. **`mod.rs`** holds the selection/streak data types and the impure dispatch (`handle_mouse`), which routes the sink and translates coordinates into the pane's own space before `pane::input::encode_mouse`. Buttons ride the same pure fn via a `Gesture`: **left** = focus the region (click-*through* to a mouse-aware child, reusing `set_pane_focus`/`vsplit_focus` so zoom-refusal and split-column restore come free), **middle** = `Effect::PasteFromClipboard` (routed through `handle_paste`, so bracketed-paste gating is inherited), **right** = `resolver.enter_leader()` + a due-now `chord_hint_due` so the which-key popup shows without the keyboard's debounce. Middle/right are never forwarded to a child.
+- **`mouse/`** — mouse routing. **`route.rs`** is the pure half: a `Copy` `MouseSnapshot` + pure `route_mouse` → `MouseSink`, plus `region_at` hit-testing the pointer against `compute_layout`'s `FrameLayout` — **the pointer, never keyboard focus**, which is what makes the wheel scroll what you're looking at. **`selection.rs`** holds the four drag-select clusters (pager text / chrome / pane text / list rows), each a `begin`+`extend`+`finish` triple where the press *claims* the drag and an unclaimed press leaves child-forwarding untouched. **`scroll.rs`** turns a wheel tick into the agent's own verified scroll keybinding for children that ignore mouse reports, escalating a sustained streak to page keys (the agent is the rate limiter, not spyc) and driving codex's `^T` overlay. **`forward.rs`** forwards to a mouse-aware child (press/release pair to the *same* child), plus the focus entry points and `mouse_report`'s coordinate translation into the pane's own space. **`mod.rs`** keeps the selection/streak data types and `handle_mouse`, the single dispatch entry. Buttons ride the same pure fn via a `Gesture`: **left** = focus the region (click-*through* to a mouse-aware child, reusing `set_pane_focus`/`vsplit_focus` so zoom-refusal and split-column restore come free), **middle** = `Effect::PasteFromClipboard` (routed through `handle_paste`, so bracketed-paste gating is inherited), **right** = `resolver.enter_leader()` + a due-now `chord_hint_due` so the which-key popup shows without the keyboard's debounce. Middle/right are never forwarded to a child.
 - **`mouse_mode.rs`** — the terminal's mouse-reporting reconcile: `settle_mouse_mode` compares `[mouse] capture` (what the user asked for) against `view.mouse_capture_on` (what the terminal is in) at loop bottom and emits `Effect::SetMouseMode` on divergence, so startup / `:mouse` / config reload / return-from-foreground all take one path.
 - Panes — **`pane_tabs.rs`** (tab lifecycle/focus), **`pane_scroll.rs`** (scroll mode + scrollback; `^a v` auto-engages an alt-screen agent's on-disk transcript), **`codex_pin.rs`** (spawn-ordered rollout claim), **`navigate.rs`**, **`quick_select.rs`** (`^a u`).
 - Vsplit — **`vsplit.rs`** (`^a |` preview cycle; `^s n`/`^s x` second Commander; `cur()`/`col(side)`; `carve_vsplit` geometry), **`preview_ops.rs`** (off-thread live-reload of the split preview).
